+++ react/src/mapper/settings/MapperSettings.tsx
@@ -1,15 +1,12 @@
-import React, { ReactNode, useCallback, useEffect, useMemo } from 'react'
+import React, { ReactNode, useCallback, useMemo } from 'react'
 
 import { articleTypes, CountsByUT } from '../../components/countsByArticleType'
 import universes_ordered from '../../data/universes_ordered'
-import { EditorError, useUndoRedo } from '../../urban-stats-script/editor-utils'
+import { EditorError } from '../../urban-stats-script/editor-utils'
 import { TypeEnvironment } from '../../urban-stats-script/types-values'
-import { Property } from '../../utils/Property'
-import { TestUtils } from '../../utils/TestUtils'
 import { settingNameStyle } from '../style'
 
 import { BetterSelector } from './BetterSelector'
-import { Selection, SelectionContext } from './SelectionContext'
 import { TopLevelEditor } from './TopLevelEditor'
 import { MapSettings } from './utils'
 
@@ -22,30 +19,6 @@ export function MapperSettings({ mapSettings, setMapSettings, errors, counts, ty
 }): ReactNode {
     const uss = mapSettings.script.uss
 
-    const selectionContext = useMemo(() => new Property<Selection | undefined>(undefined), [])
-
-    const { addState, updateCurrentSelection, ui: undoRedoUi } = useUndoRedo(
-        mapSettings,
-        selectionContext.value,
-        setMapSettings,
-        (selection) => {
-            selectionContext.value = selection
-        },
-        {
-            undoChunking: TestUtils.shared.isTesting ? 2000 : 1000,
-        },
-    )
-
-    // Update current selection when it changes
-    useEffect(() => {
-        const observer = (): void => {
-            updateCurrentSelection(selectionContext.value)
-        }
-
-        selectionContext.observers.add(observer)
-        return () => { selectionContext.observers.delete(observer) }
-    }, [selectionContext, updateCurrentSelection])
-
     const renderString = useCallback((universe: string | undefined) => ({ text: universe ?? '' }), [])
 
     const universes = useMemo(() => [undefined, ...universes_ordered], [])
@@ -54,13 +27,8 @@ export function MapperSettings({ mapSettings, setMapSettings, errors, counts, ty
         mapSettings.universe === undefined ? undefined : [undefined, ...articleTypes(counts, mapSettings.universe)] as Exclude<MapSettings['geographyKind'], undefined>[],
     [mapSettings.universe, counts])
 
-    const changeSettingsWithUndo = (newSettings: MapSettings): void => {
-        setMapSettings(newSettings)
-        addState(newSettings, selectionContext.value)
-    }
-
     return (
-        <SelectionContext.Provider value={selectionContext}>
+        <>
             <div style={settingNameStyle}>
                 Universe
             </div>
@@ -70,7 +38,7 @@ export function MapperSettings({ mapSettings, setMapSettings, errors, counts, ty
                 renderValue={renderString}
                 onChange={
                     (newUniverse) => {
-                        changeSettingsWithUndo({
+                        setMapSettings({
                             ...mapSettings,
                             universe: newUniverse,
                             geographyKind: newUniverse === undefined || mapSettings.geographyKind === undefined
@@ -93,7 +61,7 @@ export function MapperSettings({ mapSettings, setMapSettings, errors, counts, ty
                         renderValue={renderString}
                         onChange={
                             (newGeographyKind) => {
-                                changeSettingsWithUndo({
+                                setMapSettings({
                                     ...mapSettings,
                                     geographyKind: newGeographyKind,
                                 })
@@ -105,7 +73,7 @@ export function MapperSettings({ mapSettings, setMapSettings, errors, counts, ty
             <TopLevelEditor
                 uss={uss}
                 setUss={(newUss) => {
-                    changeSettingsWithUndo({
+                    setMapSettings({
                         ...mapSettings,
                         script: { uss: newUss },
                     })
@@ -113,7 +81,6 @@ export function MapperSettings({ mapSettings, setMapSettings, errors, counts, ty
                 typeEnvironment={typeEnvironment}
                 errors={errors}
             />
-            {undoRedoUi}
-        </SelectionContext.Provider>
+        </>
     )
 }
